dia4/requirements.txt
django==4.2
pillow
django-admin-tailwind
django-ckeditor